fix: fixed output errors

+++ src/services/removeFile.js
@@ -2,7 +2,8 @@ import { resolve } from 'path';
 import { unlink } from 'fs/promises';
 
 export const removeFile = async (file) => {
-  const pathToFile = resolve(file);
+  if (!file) return console.log('Invalid input');
+  const pathToFile = resolve(file); 
   try {
     await unlink(pathToFile);
     console.log(`The file has been removed`);
